test two numbers

+++ Lesson2_calculator.rb
@@ -26,7 +26,7 @@ elsif
   result=number1.to_i*number2.to_i
 elsif
   operator=='4'
-  result=number1.to_i/number2.to_i
+  result=number1.to_f/number2.to_f
 end
 
 Kernel.puts("The result is #{result}")
